move metadata into wpt-metadata and add loading state
